refactor(client):ResourceSettingsFormFields
